requirements.txt
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
python-binance==1.0.19
schedule==1.2.1
rich==13.7.0
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
argparse==1.4.0